+++ src/mbs/adapters/database/models/article.py
@@ -5,7 +5,7 @@ from sqlalchemy.orm import relationship
 
 from mbs.db import Base
 
-from src import ArticleState
+from mbs.domain.models import ArticleState
 
 
 class Article(Base):
@@ -20,11 +20,10 @@ class Article(Base):
     content = Column(VARCHAR(32000), nullable=False)
     state = Column(Enum(ArticleState), default=ArticleState.DRAFT)
     views = Column(BIGINT(), nullable=False, default=0)
-    author_id = Column(UUID(as_uuid=True), ForeignKey("users.id"), nullable=False)
+    author_id = Column(UUID(as_uuid=True), nullable=False)
     likes = relationship("models.models.like.Like", back_populates="article")
     tags = relationship('models.models.tag.Tag', secondary='article_tags', back_populates='articles')
     comments_tree = relationship("models.models.comment.CommentTree", back_populates="article")
-    files = relationship("models.models.file.File", secondary='article_files', back_populates="article")
 
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), onupdate=func.now())
@@ -54,6 +53,21 @@ class ArticleFile(Base):
 
     article_id = Column(UUID(as_uuid=True), ForeignKey("articles.id"), nullable=False)
     file_id = Column(UUID(as_uuid=True), ForeignKey("files.id"), nullable=False)
+
+    def __repr__(self):
+        return f'<{self.__class__.__name__}: {self.id}>'
+
+
+class ArticleLike(Base):
+    """
+    The ArticleLike model
+
+    """
+    __tablename__ = "article_likes"
+
+    author_id = Column(UUID(as_uuid=True), nullable=False)
+    article_id = Column(UUID(as_uuid=True), ForeignKey("articles.id"), nullable=False)
+    article = relationship("models.models.article.Article", back_populates="likes")
 
     def __repr__(self):
         return f'<{self.__class__.__name__}: {self.id}>'
